+++ lib/eventum/class.validation.php
@@ -25,66 +25,8 @@ class Validation
      * @param   string $str The string to check against
      * @return  bool
      */
-    public static function isWhitespace($str)
+    public static function isWhitespace($str): bool
     {
-        $str = trim($str);
-        if (strlen($str) == 0) {
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Method used to check whether an email address is a valid one.
-     *
-     * @param   string $str The email address to check against
-     * @return  bool
-     */
-    public static function isEmail($str)
-    {
-        $valid_chars = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i',
-                                'j', 'l', 'k', 'm', 'n', 'o', 'p', 'q', 'r',
-                                's', 't', 'u', 'w', 'v', 'x', 'y', 'z',
-                                '0', '1', '2', '3', '4', '5', '6', '7',
-                                '8', '9', ];
-        $extended_chars = ['.', '+', '_', '-', '@'];
-        $str = strtolower($str);
-
-        // we need at least one @ symbol
-        if (!strstr($str, '@')) {
-            return false;
-        }
-        // and no more than one @ symbol
-        if (strpos($str, '@') != strrpos($str, '@')) {
-            return false;
-        }
-        // check for invalid characters in the email address
-        for ($i = 0; $i < strlen($str); $i++) {
-            if ((!in_array(substr($str, $i, 1), $valid_chars)) &&
-                    (!in_array(substr($str, $i, 1), $extended_chars))) {
-                return false;
-            }
-        }
-        // email addresses need at least one dot (but also allow for user@localhost addresses)
-        if ((!strstr($str, '.')) && (substr($str, strrpos($str, '@')) != '@localhost')) {
-            return false;
-        }
-        // no two dots alongside each other
-        if (strstr($str, '..')) {
-            return false;
-        }
-        // do an extra check for a dot as the last character of an address
-        array_shift($extended_chars);
-        if ((substr($str, strlen($str) - 1) == '.') &&
-                (substr($str, strrpos($str, '@')) != '@localhost.')) {
-            return false;
-        }
-        // the last character cannot be one of the extended ones
-        if (in_array(substr($str, strlen($str) - 1), $extended_chars)) {
-            return false;
-        }
-
-        return true;
+        return trim($str) === '';
     }
 }