+++ src/Controller/Setup/SetupController.php
@@ -20,6 +20,8 @@ use Eventum\AppInfo;
 use Eventum\Controller\BaseController;
 use Eventum\Monolog\Logger;
 use Eventum\Setup\DatabaseSetup;
+use Eventum\Setup\RequirementNotSatisfiedException;
+use Eventum\Setup\Requirements;
 use Eventum\Setup\SetupException;
 use IntlCalendar;
 use Misc;
@@ -48,16 +50,15 @@ class SetupController extends BaseController
 
     protected function defaultAction()
     {
-        list($warnings, $errors) = $this->checkRequirements();
-        if ($warnings || $errors) {
-            Misc::displayRequirementErrors(array_merge($errors, $warnings), 'Eventum Setup');
-            if ($errors) {
-                exit(1);
-            }
+        try {
+            Requirements::check();
+        } catch (RequirementNotSatisfiedException $e) {
+            Misc::displayRequirementErrors($e->getErrors(), 'Eventum Setup');
+            exit(1);
         }
 
         if ($this->cat === 'install') {
-            $res = $this->install();
+            $res = $this->installAction();
             $this->tpl->assign('result', $res);
             // check for the optional IMAP extension
             $this->tpl->assign('is_imap_enabled', function_exists('imap_open'));
@@ -91,163 +92,6 @@ class SetupController extends BaseController
     protected function displayTemplate($tpl_name = null)
     {
         $this->tpl->displayTemplate(false);
-    }
-
-    /**
-     * Checks for $file for write permission.
-     *
-     * IMPORTANT: if the file does not exist, an empty file is created.
-     */
-    private function checkPermissions($file, $desc, $is_directory = false)
-    {
-        clearstatcache();
-        if (!file_exists($file)) {
-            if (!$is_directory) {
-                // try to create the file ourselves then
-                $fp = @fopen($file, 'w');
-                if (!$fp) {
-                    return $this->getPermissionError($file, $desc, $is_directory, false);
-                }
-                @fclose($fp);
-            } else {
-                if (!@mkdir($file)) {
-                    return $this->getPermissionError($file, $desc, $is_directory, false);
-                }
-            }
-        }
-        clearstatcache();
-        if (!is_writable($file)) {
-            if (stripos(PHP_OS, 'win') === false) {
-                // let's try to change the permissions ourselves
-                @chmod($file, 0644);
-                clearstatcache();
-                if (!is_writable($file)) {
-                    return $this->getPermissionError($file, $desc, $is_directory, true);
-                }
-            } else {
-                return $this->getPermissionError($file, $desc, $is_directory, true);
-            }
-        }
-        if (stripos(PHP_OS, 'win') !== false) {
-            // need to check whether we can really create files in this directory or not
-            // since is_writable() is not trustworthy on windows platforms
-            if (is_dir($file)) {
-                $fp = @fopen($file . '/dummy.txt', 'w');
-                if (!$fp) {
-                    return "$desc is not writable";
-                }
-                @fwrite($fp, 'test');
-                @fclose($fp);
-                // clean up after ourselves
-                @unlink($file . '/dummy.txt');
-            }
-        }
-
-        return '';
-    }
-
-    private function getPermissionError($file, $desc, $is_directory, $exists)
-    {
-        if ($is_directory) {
-            $title = 'Directory';
-        } else {
-            $title = 'File';
-        }
-        $error = "$title <b>'" . $file . ($is_directory ? '/' : '') . "'</b> ";
-
-        if (!$exists) {
-            $error .= "does not exist. Please create the $title and reload this page.";
-        } else {
-            $error .= "is not writeable. Please change this $title to be writeable by the web server.";
-        }
-
-        return $error;
-    }
-
-    private function checkRequirements()
-    {
-        $errors = [];
-        $warnings = [];
-
-        $extensions = [
-            // extension => array(IS_REQUIRED, MESSAGE_TO_DISPLAY)
-            'gd' => [true,
-                     'The GD extension needs to be enabled in your PHP.INI file in order for Eventum to work properly.', ],
-            'session' => [true,
-                          'The Session extension needs to be enabled in your PHP.INI file in order for Eventum to work properly.', ],
-            'pdo_mysql' => [true,
-                            'The PDO MySQL extension needs to be enabled in your PHP.INI file in order for Eventum to work properly.', ],
-            'json' => [true,
-                       'The json extension needs to be enabled in your PHP.INI file in order for Eventum to work properly.', ],
-            'mbstring' => [true,
-                           'The Multibyte String Functions extension is not enabled in your PHP installation. For localization to work properly '
-                           .
-                           'You need to install this extension. If you do not install this extension localization will be disabled.', ],
-            'iconv' => [true, 'The ICONV extension is not enabled in your PHP installation. ' .
-                        'You need to install this extension for optimal operation. If you do not install this extension some unicode data will be corrupted.', ],
-            'intl' => [false, 'The intl extension is not enabled in your PHP installation. ' .
-                        'For optional performance you should install intl extension', ],
-        ];
-
-        foreach ($extensions as $extension => $value) {
-            list($required, $message) = $value;
-            if (!extension_loaded($extension)) {
-                if ($required) {
-                    $errors[] = $message;
-                } else {
-                    $warnings[] = $message;
-                }
-            }
-        }
-
-        // check for the file_uploads php.ini directive
-        if (ini_get('file_uploads') != '1') {
-            $errors[]
-                = "The 'file_uploads' directive needs to be enabled in your PHP.INI file in order for Eventum to work properly.";
-        }
-
-        $configPath = Setup::getConfigPath();
-        $setupFile = Setup::getSetupFile();
-
-        $error = $this->checkPermissions($configPath, "Directory '" . $configPath . "'", true);
-        if (!empty($error)) {
-            $errors[] = $error;
-        }
-        $error = $this->checkPermissions($setupFile, "File '" . $setupFile . "'");
-        if (!empty($error)) {
-            $errors[] = $error;
-        }
-        $error = $this->checkPermissions(
-            $configPath . '/private_key.php', "File '" . $configPath . '/private_key.php' . "'"
-        );
-        if (!empty($error)) {
-            $errors[] = $error;
-        }
-        $error = $this->checkPermissions(
-            $configPath . '/config.php', "File '" . $configPath . '/config.php' . "'"
-        );
-        if (!empty($error)) {
-            $errors[] = $error;
-        }
-
-        $error = $this->checkPermissions(APP_LOCKS_PATH, "Directory '" . APP_LOCKS_PATH . "'", true);
-        if (!empty($error)) {
-            $errors[] = $error;
-        }
-        $error = $this->checkPermissions(APP_LOG_PATH, "Directory '" . APP_LOG_PATH . "'", true);
-        if (!empty($error)) {
-            $errors[] = $error;
-        }
-        $error = $this->checkPermissions(APP_TPL_COMPILE_PATH, "Directory '" . APP_TPL_COMPILE_PATH . "'", true);
-        if (!empty($error)) {
-            $errors[] = $error;
-        }
-        $error = $this->checkPermissions(APP_ERROR_LOG, "File '" . APP_ERROR_LOG . "'");
-        if (!empty($error)) {
-            $errors[] = $error;
-        }
-
-        return [$warnings, $errors];
     }
 
     private function getTimezone()
@@ -288,7 +132,7 @@ class SetupController extends BaseController
         return var_export($s, 1);
     }
 
-    private function initlogger()
+    private function initLogger(): void
     {
         // init timezone, logger needs it
         if (!defined('APP_DEFAULT_TIMEZONE')) {
@@ -303,9 +147,9 @@ class SetupController extends BaseController
      * return error message as string, or true indicating success
      * requires setup to be written first.
      */
-    private function setup_database()
+    private function setupDatabase(): void
     {
-        $this->initlogger();
+        $this->initLogger();
         $post = $this->getRequest()->request;
 
         $db_config = [
@@ -332,7 +176,7 @@ class SetupController extends BaseController
     /**
      * write initial values for setup file
      */
-    private function write_setup()
+    private function writeSetup(): void
     {
         $post = $this->getRequest()->request;
         $setup = $post->get('setup');
@@ -365,7 +209,7 @@ class SetupController extends BaseController
         Setup::save($setup);
     }
 
-    private function write_config()
+    private function writeConfig(): void
     {
         $post = $this->getRequest()->request;
         $configPath = Setup::getConfigPath();
@@ -395,13 +239,13 @@ class SetupController extends BaseController
         $fs->dumpFile($configFilePath, $config_contents, null);
     }
 
-    private function install()
+    private function installAction(): string
     {
         try {
             Auth::generatePrivateKey();
-            $this->write_setup();
-            $this->setup_database();
-            $this->write_config();
+            $this->writeSetup();
+            $this->setupDatabase();
+            $this->writeConfig();
         } catch (RuntimeException $e) {
             return $e->getMessage();
         }